actualizando proyecto

+++ src/main/java/com/network/Pilotosformulauno/Servicio/IServicio_Temporada.java
@@ -2,6 +2,6 @@ package com.network.Pilotosformulauno.Servicio;
 
 import com.network.Pilotosformulauno.Dominio.Temporada;
 
-public interface IServicio_Temporada extends ICrud<Temporada,Integer> {
+public interface IServicio_Temporada extends ICrud<Temporada,Long> {
     ;
 }
